src/fallbacks.py
```python
from src.user_profile import user_profile
from src.integration.uber_eats_api import fetch_menu, place_order
from src.utils.exceptions import DishUnavailableError

def pick_alternative(dish, cuisine):
    """
    Function to pick an alternative dish if the preferred dish isn't available.
    """
    try:
        # Fetch the menu
        menu = fetch_menu(cuisine)

        # If the preferred dish is not in the menu, pick an alternative
        if dish not in menu:
            alternative_dish = menu[0]  # Pick the first dish as an alternative
            return alternative_dish
        else:
            return dish

    except DishUnavailableError:
        print("The preferred dish is not available. An alternative dish has been selected.")

def skip_meal_or_alternative_service():
    """
    Function to skip meals or opt for a different delivery service in case Uber Eats is unavailable.
    """
    # Check if Uber Eats is available
    if not uber_eats_available():
        # If not, skip the meal or opt for a different delivery service
        print("Uber Eats is not available. Skipping the meal or opting for a different delivery service.")
        return None

    # If Uber Eats is available, proceed with the order
    else:
        return place_order(user_profile["preferred_dish"], user_profile["preferred_cuisine"])
```